+++ web/src/dictionary/markup/types.ts
